+++ src/hooks/login/index.ts
@@ -6,3 +6,4 @@ export { useGetModalLoginMethods } from './useGetModalLoginMethods';
 export { useLedgerLogin } from './useLedgerLogin';
 export { useXaliasLogin } from './useXaliasLogin';
 export { useWalletConnectV2Login } from './useWalletConnectV2Login';
+export { useMetamaskLogin } from './useMetamaskLogin';