+++ src/hooks/transactions/useSignTransactions.tsx
@@ -6,7 +6,9 @@ import {
 } from '@multiversx/sdk-core';
 
 import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
-import { CrossWindowProvider } from '@multiversx/sdk-web-wallet-cross-window-provider';
+import { MetamaskProvider } from '@multiversx/sdk-metamask-provider/out/metamaskProvider';
+import { CrossWindowProvider } from '@multiversx/sdk-web-wallet-cross-window-provider/out/CrossWindowProvider/CrossWindowProvider';
+
 import uniq from 'lodash/uniq';
 import {
   ERROR_SIGNING,
@@ -83,6 +85,7 @@ export const useSignTransactions = () => {
   function clearSignInfo(sessionId?: string) {
     const isExtensionProvider = provider instanceof ExtensionProvider;
     const isCrossWindowProvider = provider instanceof CrossWindowProvider;
+    const isMetamaskProvider = provider instanceof MetamaskProvider;
     const isExperiementalWebviewProvider =
       provider instanceof ExperimentalWebviewProvider;
 
@@ -91,7 +94,7 @@ export const useSignTransactions = () => {
 
     isSigningRef.current = false;
 
-    if (!isExtensionProvider && !isCrossWindowProvider) {
+    if (!isExtensionProvider && !isCrossWindowProvider && !isMetamaskProvider) {
       return;
     }
 
@@ -99,6 +102,9 @@ export const useSignTransactions = () => {
 
     if (isExtensionProvider) {
       ExtensionProvider.getInstance()?.cancelAction?.();
+    }
+    if (isMetamaskProvider) {
+      MetamaskProvider.getInstance()?.cancelAction?.();
     }
     if (isCrossWindowProvider) {
       CrossWindowProvider.getInstance()?.cancelAction?.();
@@ -191,9 +197,16 @@ export const useSignTransactions = () => {
     }
 
     const allowGuardian = !customTransactionInformation.skipGuardian;
+    const hasConsentPopup = customTransactionInformation.hasConsentPopup;
+    const isCrossWindowProvider = provider instanceof CrossWindowProvider;
 
     try {
       isSigningRef.current = true;
+
+      if (isCrossWindowProvider && hasConsentPopup) {
+        (provider as CrossWindowProvider).setShouldShowConsentPopup(true);
+      }
+
       const signedTransactions: Transaction[] =
         (await provider.signTransactions(
           isGuarded && allowGuardian