+++ src/providers/accountProvider.ts
@@ -1,7 +1,8 @@
 import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
 import { HWProvider } from '@multiversx/sdk-hw-provider';
+import { MetamaskProvider } from '@multiversx/sdk-metamask-provider/out/metamaskProvider';
 import { OperaProvider } from '@multiversx/sdk-opera-provider';
-import { CrossWindowProvider } from '@multiversx/sdk-web-wallet-cross-window-provider';
+import { CrossWindowProvider } from '@multiversx/sdk-web-wallet-cross-window-provider/out/CrossWindowProvider/CrossWindowProvider';
 import { WalletProvider } from '@multiversx/sdk-web-wallet-provider';
 import { IDappProvider } from 'types';
 import { WalletConnectV2Provider } from 'utils/walletconnect/__sdkWalletconnectProvider';
@@ -10,6 +11,7 @@ import { emptyProvider } from './utils';
 export type ProvidersType =
   | IDappProvider
   | ExtensionProvider
+  | MetamaskProvider
   | WalletProvider
   | CrossWindowProvider
   | HWProvider