+++ src/utils/platform/getIsMetamaskAvailable.ts
@@ -1,0 +1,5 @@
+import { MetamaskProvider } from '@multiversx/sdk-metamask-provider/out/metamaskProvider';
+
+export const getIsMetamaskAvailable = () => {
+  return MetamaskProvider.isMetamaskInstalled();
+};